Combine PlayerMetadata collection to UserDocument and rename it to PlayerAccount
